Refactor test class and orgnize as partial class.
